+++ dist_node/lex/punctuator_lexer.js
@@ -5,10 +5,11 @@
 var __o = require("bennu")["parse"],
     __o0 = require("bennu")["text"],
     __o1 = require("nu-stream")["stream"],
-    punctuator, operator, always = __o["always"],
+    punctuator, prefixOperator, infixOperator, operator, always = __o["always"],
     binds = __o["binds"],
     fail = __o["fail"],
     label = __o["label"],
+    either = __o["either"],
     enumeration = __o["enumeration"],
     map = __o["map"],
     many = __o["many"],
@@ -23,15 +24,23 @@ var __o = require("bennu")["parse"],
         "§"
     ];
 (punctuator = label("Punctuator Lexer", trie(punctuators)));
-var operators = ["&&", "||", "<<", ">>", ">>>", "<=", "<", ">=", ">", "===", "!==", "==", "!=", "!", "~", "++", "--",
-    "&", "|", "^", "+", "-", "*", "/", "%", "|>", "<|", "\\>", "\\>>", "<\\", "<<\\"
-],
-    operatorChar = oneOf("?+-*/%|&^<>=!~@");
-(operator = label("Operator Lexer", binds(enumeration(trie(operators), join(many(operatorChar))), (function() {
+var operatorChar = oneOf("?+-*/%|&^<>=!~@");
+(prefixOperator = label("Prefix Operator Lexer", binds(enumeration(trie(["~", "!", "++", "--"]), join(many(operatorChar))), (
+    function() {
+        var args = arguments;
+        return always(__add.apply(null, args));
+    }))));
+(infixOperator = label("Infix Operator Lexer", binds(enumeration(trie(["&&", "||", "<<", ">>", ">>>", "<=", "<", ">=",
+    ">", "===", "!==", "==", "!=", "&", "|", "^", "+", "-", "*", "/", "%", "|>", "<|", "\\>",
+    "\\>>", "<\\", "<<\\"
+]), join(many(operatorChar))), (function() {
     var args = arguments,
         x = __add.apply(null, args);
     return ((punctuators.indexOf(x) >= 0) ? fail((("Cannot use puctuator " + x) + " as operator")) :
         always(x));
 }))));
+(operator = label("Prefix Operator Lexer", either(prefixOperator, infixOperator)));
 (exports["punctuator"] = punctuator);
+(exports["prefixOperator"] = prefixOperator);
+(exports["infixOperator"] = infixOperator);
 (exports["operator"] = operator);