+++ dist_node/parse/package_parser.js
@@ -1,15 +1,10 @@
 /*
- * THIS FILE IS AUTO GENERATED from 'lib/parse/package_parser.kep'
+ * THIS FILE IS AUTO GENERATED FROM 'lib/parse/package_parser.kep'
  * DO NOT EDIT
-*/"use strict";
+*/
+"use strict";
 var __o = require("bennu")["parse"],
-    __o0 = require("bennu")["lang"],
-    ast_package = require("khepri-ast")["package"],
-    __o1 = require("./common"),
-    __o2 = require("./token_parser"),
-    __o3 = require("./value_parser"),
-    __o4 = require("./statement_parser"),
-    khepriPackage, attempt = __o["attempt"],
+    attempt = __o["attempt"],
     eager = __o["eager"],
     either = __o["either"],
     enumeration = __o["enumeration"],
@@ -17,18 +12,25 @@ var __o = require("bennu")["parse"],
     next = __o["next"],
     optional = __o["optional"],
     label = __o["label"],
+    late = __o["late"],
+    __o0 = require("bennu")["lang"],
     between = __o0["between"],
     sepBy = __o0["sepBy"],
+    ast_package = require("khepri-ast")["package"],
+    __o1 = require("./common"),
     node = __o1["node"],
     nodea = __o1["nodea"],
+    __o2 = require("./token_parser"),
     keyword = __o2["keyword"],
     punctuator = __o2["punctuator"],
+    __o3 = require("./value_parser"),
     identifier = __o3["identifier"],
     stringLiteral = __o3["stringLiteral"],
     operator = __o3["operator"],
+    __o4 = require("./statement_parser"),
     withStatement = __o4["withStatement"],
     blockStatement = __o4["blockStatement"],
-    symbol = either(identifier, operator),
+    khepriPackage, symbol = either(identifier, operator),
     packageExport = label("Package Export", either(node(symbol, ast_package.PackageExport.create), nodea(enumeration(
         stringLiteral, next(punctuator(":"), identifier)), (function(loc, alias, id) {
         return ast_package.PackageExport.create(loc, id, alias);