+++ dist_node/parse/value_parser.js
@@ -1,16 +1,19 @@
 /*
- * THIS FILE IS AUTO GENERATED from 'lib/parse/value_parser.kep'
+ * THIS FILE IS AUTO GENERATED FROM 'lib/parse/value_parser.kep'
  * DO NOT EDIT
-*/"use strict";
+*/
+"use strict";
 var __o = require("bennu")["parse"],
+    bind = __o["bind"],
+    choice = __o["choice"],
+    label = __o["label"],
     __o0 = require("bennu")["lang"],
+    between = __o0["between"],
     ast_value = require("khepri-ast")["value"],
     token = require("./token_parser"),
+    prefixedOp = token["prefixedOp"],
     literal, nullLiteral, booleanLiteral, numericLiteral, stringLiteral, regularExpressionLiteral, identifier, operator,
-        unaryOperator, binaryOperator, choice = __o["choice"],
-    label = __o["label"],
-    between = __o0["between"],
-    p;
+        unaryOperator, binaryOperator, p;
 (nullLiteral = label("Null Literal", ((p = token.nullLiteral), p.map((function(x) {
     return ast_value.Literal.create(x.loc, "null", x.value);
 })))));